# swagger.json
{
	"components": {
...
		"schemas": {
...
			"EProfileStatus": {
				"enum": [
					"online",
					"offline"
				],
				"type": "string"
			},
...
			"PublicProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EProfileStatus"
					},
					"lastOnline": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BotCommandDto": {
				"properties": {
					"command": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"command",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BotDetailDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"token": {
						"type": "string"
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true
					},
					"commands": {
						"items": {
							"$ref": "#/components/schemas/BotCommandDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"username",
					"displayName",
					"description",
					"avatarUrl",
					"isActive",
					"createdAt",
					"token",
					"webhookUrl",
					"commands"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateBotBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"displayName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BotDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"displayName",
					"description",
					"avatarUrl",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateBotBody": {
				"properties": {
					"displayName": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"avatarId": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISetWebhookBody": {
				"properties": {
					"url": {
						"type": "string"
					},
					"secret": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISetCommandsBody": {
				"properties": {
					"commands": {
						"items": {
							"properties": {
								"description": {
									"type": "string"
								},
								"command": {
									"type": "string"
								}
							},
							"required": [
								"description",
								"command"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"commands"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ECallType": {
				"enum": [
					"voice",
					"video"
				],
				"type": "string"
			},
			"ECallStatus": {
				"enum": [
					"ringing",
					"active",
					"ended",
					"missed",
					"declined"
				],
				"type": "string"
			},
			"CallDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"callerId": {
						"type": "string"
					},
					"calleeId": {
						"type": "string"
					},
					"chatId": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"$ref": "#/components/schemas/ECallType"
					},
					"status": {
						"$ref": "#/components/schemas/ECallStatus"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"endedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"duration": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"caller": {
						"properties": {
							"avatarUrl": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"callee": {
						"properties": {
							"avatarUrl": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"callerId",
					"calleeId",
					"chatId",
					"type",
					"status",
					"startedAt",
					"endedAt",
					"duration",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IInitiateCallBody": {
				"properties": {
					"calleeId": {
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ECallType"
					}
				},
				"required": [
					"calleeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICallHistoryDto": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/CallDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISetSlowModeBody": {
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"seconds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBanMemberBody": {
				"properties": {
					"duration": {
						"type": "number",
						"format": "double"
					},
					"reason": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EChatType": {
				"enum": [
					"direct",
					"group",
					"channel",
					"secret"
				],
				"type": "string"
			},
			"EChatMemberRole": {
				"enum": [
					"owner",
					"admin",
					"member",
					"subscriber"
				],
				"type": "string"
			},
			"ChatMemberDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/EChatMemberRole"
					},
					"joinedAt": {
						"type": "string",
						"format": "date-time"
					},
					"mutedUntil": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"isPinnedChat": {
						"type": "boolean"
					},
					"isArchived": {
						"type": "boolean"
					},
					"folderId": {
						"type": "string",
						"nullable": true
					},
					"profile": {
						"$ref": "#/components/schemas/PublicProfileDto"
					}
				},
				"required": [
					"id",
					"userId",
					"role",
					"joinedAt",
					"mutedUntil",
					"isPinnedChat",
					"isArchived",
					"folderId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChatDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EChatType"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"isPublic": {
						"type": "boolean"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdById": {
						"type": "string"
					},
					"lastMessageAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/ChatMemberDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"type",
					"name",
					"description",
					"username",
					"isPublic",
					"avatarUrl",
					"createdById",
					"lastMessageAt",
					"createdAt",
					"updatedAt",
					"members"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IChatListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ChatDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.1",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/bot": {
			"post": {
				"operationId": "CreateBot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BotDetailDto"
								}
							}
						}
					}
				},
				"summary": "Создать бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateBotBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMyBots",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BotDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Мои боты",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/bot/{id}": {
			"get": {
				"operationId": "GetBotById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BotDetailDto"
								}
							}
						}
					}
				},
				"summary": "Детали бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BotDetailDto"
								}
							}
						}
					}
				},
				"summary": "Обновить бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateBotBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBot",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Удалить бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bot/{id}/token": {
			"post": {
				"operationId": "RegenerateToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BotDetailDto"
								}
							}
						}
					}
				},
				"summary": "Перегенерировать токен",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bot/{id}/webhook": {
			"post": {
				"operationId": "SetWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BotDetailDto"
								}
							}
						}
					}
				},
				"summary": "Установить webhook",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetWebhookBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWebhook",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Удалить webhook",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/bot/{id}/commands": {
			"post": {
				"operationId": "SetCommands",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BotCommandDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Установить команды бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetCommandsBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCommands",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BotCommandDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Получить команды бота",
				"tags": [
					"Bot"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/call": {
			"post": {
				"operationId": "InitiateCall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CallDto"
								}
							}
						}
					}
				},
				"description": "Инициировать звонок.",
				"summary": "Начать звонок",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IInitiateCallBody"
							}
						}
					}
				}
			}
		},
		"/api/call/{id}/answer": {
			"post": {
				"operationId": "AnswerCall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CallDto"
								}
							}
						}
					}
				},
				"description": "Ответить на звонок.",
				"summary": "Ответить",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/call/{id}/decline": {
			"post": {
				"operationId": "DeclineCall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CallDto"
								}
							}
						}
					}
				},
				"description": "Отклонить звонок.",
				"summary": "Отклонить",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/call/{id}/end": {
			"post": {
				"operationId": "EndCall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CallDto"
								}
							}
						}
					}
				},
				"description": "Завершить звонок.",
				"summary": "Завершить",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/call/history": {
			"get": {
				"operationId": "GetCallHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICallHistoryDto"
								}
							}
						}
					}
				},
				"description": "Получить историю звонков.",
				"summary": "История звонков",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/call/active": {
			"get": {
				"operationId": "GetActiveCall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CallDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Получить активный звонок.",
				"summary": "Активный звонок",
				"tags": [
					"Call"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/chat/{id}/slow-mode": {
			"patch": {
				"operationId": "SetSlowMode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"slowModeSeconds": {
											"type": "number",
											"format": "double"
										},
										"chatId": {
											"type": "string"
										}
									},
									"required": [
										"slowModeSeconds",
										"chatId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Установить режим медленной отправки.",
				"summary": "Медленный режим",
				"tags": [
					"Chat Moderation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetSlowModeBody"
							}
						}
					}
				}
			}
		},
		"/api/chat/{id}/members/{userId}/ban": {
			"post": {
				"operationId": "BanMember",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Заблокировать участника чата.",
				"summary": "Блокировка участника",
				"tags": [
					"Chat Moderation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBanMemberBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "UnbanMember",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Разблокировать участника чата.",
				"summary": "Разблокировка участника",
				"tags": [
					"Chat Moderation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/chat/{id}/members/banned": {
			"get": {
				"operationId": "GetBannedMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Получить заблокированных участников.",
				"summary": "Заблокированные участники",
				"tags": [
					"Chat Moderation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/chat": {
			"get": {
				"operationId": "GetUserChats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IChatListDto"
								}
							}
						}
					}
				},
				"description": "Получить список чатов текущего пользователя.",
				"summary": "Список чатов",
				"tags": [
					"Chat"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		},
		{
			"url": "http://0.0.0.0:8181",
			"description": "Local dev"
		},
		{
			"url": "http://147.45.245.104:8181",
			"description": "Host"
		}
	]
}